mcp-agents/requirements.txt
httpx>=0.27.0
pyyaml>=6.0
mcp>=1.0.0
